requirements.txt
gitpython>=3.1.40
requests>=2.31.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
packaging>=23.2
bandit>=1.7.5
safety>=2.3.5
openai>=1.3.0
anthropic>=0.7.0
